refactor: move notification logic into /events/process AB#33720
