+++ services/API-service/src/api/event/event.controller.ts
@@ -22,7 +22,6 @@ import { AlertArea, EventSummaryCountry } from '../../shared/data.model';
 import { DisasterType } from '../disaster-type/disaster-type.enum';
 import { UserRole } from '../user/user-role.enum';
 import { UserDecorator } from '../user/user.decorator';
-import { CountryDisasterTypeDto } from './dto/country-disaster-type.dto';
 import {
   ActivationLogDto,
   EventPlaceCodeDto,
@@ -219,47 +218,5 @@ export class EventController {
     @Body() uploadAlertPerLeadTimeDto: UploadAlertPerLeadTimeDto,
   ): Promise<void> {
     await this.eventService.uploadAlertPerLeadTime(uploadAlertPerLeadTimeDto);
-  }
-
-  // NOTE: keep this endpoint in until all pipelines migrated to /event/process
-  @UseGuards(RolesGuard)
-  @Roles(UserRole.PipelineUser)
-  @ApiOperation({
-    summary:
-      'Close events automatically for given country and disaster-type. Must be run at end of every pipeline. As a backup, the same logic is also in /notification/send endpoint.',
-  })
-  @ApiResponse({
-    status: 201,
-    description: 'Closed finished events.',
-  })
-  @Post('close-events')
-  public async closeEvents(
-    @Body() closeEventsDto: CountryDisasterTypeDto,
-  ): Promise<void> {
-    // NOTE: this old endpoint will also point to this new method already
-    await this.eventService.processEvents(
-      closeEventsDto.countryCodeISO3,
-      closeEventsDto.disasterType,
-    );
-  }
-
-  @UseGuards(RolesGuard)
-  @Roles(UserRole.PipelineUser)
-  @ApiOperation({
-    summary:
-      'Process events for given country and disaster-type. Must be run at end of every pipeline.',
-  })
-  @ApiResponse({
-    status: 201,
-    description: 'Processed events.',
-  })
-  @Post('process')
-  public async processEvents(
-    @Body() processEventsDto: CountryDisasterTypeDto,
-  ): Promise<void> {
-    await this.eventService.processEvents(
-      processEventsDto.countryCodeISO3,
-      processEventsDto.disasterType,
-    );
   }
 }
