+++ services/API-service/src/api/notification/notification-content/notification-content.module.ts
@@ -3,7 +3,6 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { HelperService } from '../../../shared/helper.service';
 import { AdminAreaDataModule } from '../../admin-area-data/admin-area-data.module';
-import { AdminAreaDynamicDataModule } from '../../admin-area-dynamic-data/admin-area-dynamic-data.module';
 import { AdminAreaModule } from '../../admin-area/admin-area.module';
 import { CountryEntity } from '../../country/country.entity';
 import { DisasterTypeEntity } from '../../disaster-type/disaster-type.entity';
@@ -19,7 +18,6 @@ import { NotificationContentService } from './notification-content.service';
       DisasterTypeEntity,
     ]),
     EventModule,
-    AdminAreaDynamicDataModule,
     AdminAreaDataModule,
     AdminAreaModule,
   ],
