+++ services/API-service/src/app.module.ts
@@ -13,6 +13,7 @@ import { MetadataModule } from './api/metadata/metadata.module';
 import { NotificationModule } from './api/notification/notification.module';
 import { WhatsappModule } from './api/notification/whatsapp/whatsapp.module';
 import { PointDataModule } from './api/point-data/point-data.module';
+import { ProcessPipelineModule } from './api/process-pipeline/process-pipeline.module';
 import { TyphoonTrackModule } from './api/typhoon-track/typhoon-track.module';
 import { UserModule } from './api/user/user.module';
 import { WaterpointsModule } from './api/waterpoints/waterpoints.module';
@@ -31,6 +32,7 @@ import { TypeOrmModule } from './typeorm.module';
     HealthModule,
     CountryModule,
     EventModule,
+    ProcessPipelineModule,
     MetadataModule,
     AdminAreaModule,
     AdminAreaDynamicDataModule,
